+++ Jenkinsfile
@@ -26,80 +26,26 @@ def packageSpecificExtensions() {
 	}
 }
 
-def publishPreReleaseExtensions() {
-	stage "replace extension version"
-	sh "node ./scripts/prepare-nightly-build.js"
-	sh "mv ./package.insiders.json ./package.json"
-
-	def packageJson = readJSON file: 'package.json'
-	env.EXTENSION_VERSION = "${packageJson.version}"
-
-	stage "publish generic version"
-	withCredentials([[$class: 'StringBinding', credentialsId: 'vscode_java_marketplace', variable: 'TOKEN']]) {
-		sh 'vsce publish --pre-release -p ${TOKEN} --target win32-ia32 win32-arm64 linux-armhf alpine-x64 alpine-arm64'
-	}
-
-	stage "publish specific version"
-	packageSpecificExtensions()
-	withCredentials([[$class: 'StringBinding', credentialsId: 'vscode_java_marketplace', variable: 'TOKEN']]) {
-		def platformVsixes = findFiles(glob: '**.vsix')
-		for(platformVsix in platformVsixes){
-			sh 'vsce publish -p ${TOKEN}' + " --packagePath ${platformVsix.path}"
-		}
-	}
-}
-
-node('rhel8'){
-	stage 'Build JDT LS'
-
-	env.JAVA_HOME="${tool 'openjdk-11'}"
-	env.PATH="${env.JAVA_HOME}/bin:${env.PATH}"
-	sh 'java -version'
-
-	git url: 'https://github.com/eclipse/eclipse.jdt.ls.git'
-	sh "./mvnw clean verify -B -U -e -Pserver-distro -Dtycho.disableP2Mirrors=true -DskipTests -P!jboss-maven-repos,!redhat-ga-repository,!redhat-ea-repository"
-
-	def files = findFiles(glob: '**/org.eclipse.jdt.ls.product/distro/**.tar.gz')
-	stash name: 'server_distro', includes :files[0].path
-}
-
-node('rhel8'){
-	env.JAVA_HOME="${tool 'openjdk-11'}"
-	env.PATH="${env.JAVA_HOME}/bin:${env.PATH}"
-	stage 'Checkout vscode-java code'
-	deleteDir()
-	git url: 'https://github.com/redhat-developer/vscode-java.git'
-
-	stage 'install vscode-java build requirements'
-	installBuildRequirements()
-
-	stage 'Build vscode-java'
-	buildVscodeExtension()
-	unstash 'server_distro'
-	def files = findFiles(glob: '**/org.eclipse.jdt.ls.product/distro/**.tar.gz')
-	sh "rm -rf ./out"
-	sh "mkdir ./server"
-	sh "tar -xvzf ${files[0].path} -C ./server"
-
+def packageExtensions() {
 	if (publishPreRelease.equals('true')) {
-		publishPreReleaseExtensions()
+		// for pre-release versions we only package platform specific extensions
+		stage "replace extension version"
+		sh "npx gulp prepare_pre_release"
+
+		def packageJson = readJSON file: 'package.json'
+		env.EXTENSION_VERSION = "${packageJson.version}"
+
+		packageSpecificExtensions()
 	} else {
-		stage "Package vscode-java"
+		stage "package generic version"
 		def packageJson = readJSON file: 'package.json'
 		env.EXTENSION_VERSION = "${packageJson.version}"
 
 		sh "vsce package -o java-${env.EXTENSION_VERSION}-${env.BUILD_NUMBER}.vsix"
 
-		stage 'Test vscode-java for staging'
-		wrap([$class: 'Xvnc']) {
-			sh "npm run compile" //compile the test code too
-			env.SKIP_COMMANDS_TEST="true"
-			sh "npm test --silent"
-		}
 		def vsix = findFiles(glob: '**.vsix')
 		stash name:'vsix', includes:vsix[0].path
 
-		// Package platform specific versions
 		packageSpecificExtensions()
 		stash name:'platformVsix', includes:'java-win32-*.vsix,java-linux-*.vsix,java-darwin-*.vsix'
 
@@ -114,8 +60,22 @@ node('rhel8'){
 	}
 }
 
-node('rhel8'){
-	if(publishToMarketPlace.equals('true')){
+def publishExtensions() {
+	if (publishPreRelease.equals('true')) {
+		stage "publish generic version"
+		withCredentials([[$class: 'StringBinding', credentialsId: 'vscode_java_marketplace', variable: 'TOKEN']]) {
+			sh 'vsce publish --pre-release -p ${TOKEN} --target win32-ia32 win32-arm64 linux-armhf alpine-x64 alpine-arm64'
+		}
+
+		stage "publish specific version"
+		// for pre-release versions, vsixs are not stashed and kept in project folder
+		withCredentials([[$class: 'StringBinding', credentialsId: 'vscode_java_marketplace', variable: 'TOKEN']]) {
+			def platformVsixes = findFiles(glob: '**.vsix')
+			for(platformVsix in platformVsixes){
+				sh 'vsce publish -p ${TOKEN}' + " --packagePath ${platformVsix.path}"
+			}
+		}
+	} else if (publishToMarketPlace.equals('true')) {
 		timeout(time:5, unit:'DAYS') {
 			input message:'Approve deployment?', submitter: 'fbricon,rgrunber'
 		}
@@ -154,5 +114,49 @@ node('rhel8'){
 
 		// copy this stable build to Akamai-mirrored /static/ URL, so staging can be cleaned out more easily
 		sh "sftp ${UPLOAD_LOCATION}/static/jdt.ls/stable/ <<< \$'mkdir ${artifactDir}\nput -r ${artifactDir}'"
-	}// if publishToMarketPlace
+	}
+}
+
+node('rhel8'){
+	stage 'Build JDT LS'
+
+	env.JAVA_HOME="${tool 'openjdk-11'}"
+	env.PATH="${env.JAVA_HOME}/bin:${env.PATH}"
+	sh 'java -version'
+
+	git url: 'https://github.com/eclipse/eclipse.jdt.ls.git'
+	sh "./mvnw clean verify -B -U -e -Pserver-distro -Dtycho.disableP2Mirrors=true -DskipTests -P!jboss-maven-repos,!redhat-ga-repository,!redhat-ea-repository"
+
+	def files = findFiles(glob: '**/org.eclipse.jdt.ls.product/distro/**.tar.gz')
+	stash name: 'server_distro', includes :files[0].path
+}
+
+node('rhel8'){
+	env.JAVA_HOME="${tool 'openjdk-11'}"
+	env.PATH="${env.JAVA_HOME}/bin:${env.PATH}"
+	stage 'Checkout vscode-java code'
+	deleteDir()
+	git url: 'https://github.com/redhat-developer/vscode-java.git'
+
+	stage 'install vscode-java build requirements'
+	installBuildRequirements()
+
+	stage 'Build vscode-java'
+	buildVscodeExtension()
+	unstash 'server_distro'
+	def files = findFiles(glob: '**/org.eclipse.jdt.ls.product/distro/**.tar.gz')
+	sh "rm -rf ./out"
+	sh "mkdir ./server"
+	sh "tar -xvzf ${files[0].path} -C ./server"
+
+	stage 'Test vscode-java for staging'
+	wrap([$class: 'Xvnc']) {
+		sh "npm run compile" //compile the test code too
+		env.SKIP_COMMANDS_TEST="true"
+		sh "npm test --silent"
+	}
+
+	packageExtensions()
+
+	publishExtensions()
 }
