+++ gulpfile.js
@@ -177,6 +177,23 @@ gulp.task('clean_test_folder', function(done) {
 	done();
 });
 
+gulp.task('prepare_pre_release', function(done) {
+	const json = JSON.parse(fse.readFileSync("./package.json").toString());
+	const stableVersion = json.version.match(/(\d+)\.(\d+)\.(\d+)/);
+	const major = stableVersion[1];
+	const minor = stableVersion[2];
+	const date = new Date();
+	const month = date.getMonth() + 1;
+	const day = date.getDate();
+	const hours = date.getHours();
+	const patch = `${date.getFullYear()}${prependZero(month)}${prependZero(day)}${prependZero(hours)}`;
+	const insiderPackageJson = Object.assign(json, {
+		version: `${major}.${minor}.${patch}`,
+	});
+	fse.writeFileSync("./package.json", JSON.stringify(insiderPackageJson, null, 2));
+	done();
+});
+
 function isWin() {
 	return /^win/.test(process.platform);
 }
@@ -191,4 +208,11 @@ function isLinux() {
 
 function mvnw() {
 	return isWin()?"mvnw.cmd":"./mvnw";
+}
+
+function prependZero(number) {
+	if (number > 99) {
+		throw "Unexpected value to prepend with zero";
+	}
+	return `${number < 10 ? "0" : ""}${number}`;
 }
